requirements.txt
pytesseract
Pillow
pyperclip
pystray
keyboard
